@@ -9,25 +9,25 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "ext-gd": "*",
     "ext-json": "*",
     "ext-zlib": "*",
-    "phing/phing": "^3.0.0-RC3",
-    "plaisio/helper-url": "^1.1.2",
-    "plaisio/kernel": "^3.1.1",
-    "plaisio/web-assets": "^3.0.0",
-    "setbased/exception": "^2.2.0",
-    "setbased/helper-cast": "^2.1.1",
-    "setbased/helper-program-execution": "^1.1.0",
-    "setbased/php-stratum-sqlite-pdo": "^1.5.1",
-    "symfony/filesystem": "^5.4"
+    "phing/phing": "^3.0.0-RC4",
+    "plaisio/helper-url": "^1.2.0",
+    "plaisio/kernel": "^3.2.1",
+    "plaisio/web-assets": "^3.1.0",
+    "setbased/exception": "^2.3.0",
+    "setbased/helper-cast": "^3.0.0",
+    "setbased/helper-program-execution": "^1.2.0",
+    "setbased/php-stratum-sqlite-pdo": "^1.7.0",
+    "symfony/filesystem": "^6.2.5"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.5.2",
-    "plaisio/console": "^2.4.8",
-    "plaisio/page-core": "^1.4.2",
-    "setbased/php-stratum": "^6.1.0"
+    "phpunit/phpunit": "^9.6.3",
+    "plaisio/console-kernel": "^1.0.1",
+    "plaisio/page-core": "^1.5.0",
+    "setbased/php-stratum": "^6.3.1"
   },
   "autoload": {
     "psr-4": {
